nodejs postress practice

+++ node-express-tutorial/src/usingDB/controllers/Reflection.js
@@ -5,22 +5,23 @@ import db from '../db';
 const Reflection = {
 	async create(req, res) {
 		const
-				text = `
+				createQuery = `
 					INSERT INTO
-					reflections(id, success, low_point, take_away, created_date, modified_date)
-					VALUES($1, $2, $3, $4, $5, $6)
+					reflections(id, success, low_point, take_away, owner_id, created_date, modified_date)
+					VALUES($1, $2, $3, $4, $5, $6, $7)
 					returning *`,
 				values = [
 						uuidv4(),
 						req.body.success,
 						req.body.low_point,
 						req.body.take_away,
+						req.user.id,
 						moment(new Date()),
 						moment(new Date())
 				];
 
 		try {
-			const { rows } = await db.query(text, values);
+			const { rows } = await db.query(createQuery, values);
 
 			return res.status(201).send(rows[0]);
 		} catch (error) {
@@ -29,10 +30,10 @@ const Reflection = {
 	},
 
 	async getAll(req, res) {
-		const findAllQuery = 'SELECT * FROM reflections';
+		const findAllQuery = 'SELECT * FROM reflections WHERE owner_id = $1';
 
 		try {
-			const { rows, rowCount } = await db.query(findAllQuery);
+			const { rows, rowCount } = await db.query(findAllQuery, [req.user.id]);
 
 			return res.status(200).send({ rows, rowCount });
 		} catch (error) {
@@ -41,10 +42,10 @@ const Reflection = {
 	},
 
 	async getOne(req, res) {
-		const text = 'SELECT * FROM reflections WHERE id = $1';
+		const text = 'SELECT * FROM reflections WHERE id = $1 AND owner_id = $2';
 
 		try {
-			const { rows } = await db.query(text, [req.params.id]);
+			const { rows } = await db.query(text, [req.params.id, req.user.id]);
 			if (!rows[0]) {
 				return res.status(404).send({'message': 'reflection not found'});
 			}
@@ -57,14 +58,14 @@ const Reflection = {
 
 	async update(req, res) {
 		const
-				findOneQuery = 'SELECT * FROM reflections WHERE id = $1',
+				findOneQuery = 'SELECT * FROM reflections WHERE id = $1 AND owner_id = $2',
 				updateOneQuery = `
 					UPDATE reflections
 					SET success=$1,low_point=$2,take_away=$3,modified_date=$4
-					WHERE id=$5 returning *`;
+					WHERE id=$5 AND owner_id=$6 returning *`;
 
 		try {
-			const { rows } = await db.query(findOneQuery, [req.params.id]);
+			const { rows } = await db.query(findOneQuery, [req.params.id, req.user.id]);
 			if (!rows[0]) {
 				return res.status(404).send({'message': 'reflection not found'});
 			}
@@ -74,7 +75,8 @@ const Reflection = {
 					req.body.low_point || rows[0].low_point,
 					req.body.take_away || rows[0].take_away,
 					moment(new Date()),
-					req.params.id
+					req.params.id,
+					req.user.id
 			];
 			const response = await db.query(updateOneQuery, values);
 
@@ -85,10 +87,10 @@ const Reflection = {
 	},
 
 	async delete(req, res) {
-		const deleteQuery = 'DELETE FROM reflections WHERE id=$1 returning *';
+		const deleteQuery = 'DELETE FROM reflections WHERE id=$1 AND owner_id=$2 returning *';
 
 		try {
-			const { rows } = await db.query(deleteQuery, [req.params.id]);
+			const { rows } = await db.query(deleteQuery, [req.params.id, req.user.id]);
 			if (!rows[0]) {
 				return res.status(404).send({'message': 'reflection not found'});
 			}
